requirements.txt
# TensorFlow and TensorFlow Lite dependencies
tensorflow==2.15.0
tensorflow-hub==0.15.0
tensorboard==2.15.0

# Data processing and visualization
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.1
opencv-python==4.8.1.78

# Machine learning utilities
scikit-learn==1.3.0
pandas==2.0.3

# Hyperparameter tuning
keras-tuner==1.4.5

# Model optimization
tensorflow-model-optimization==0.8.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1 